+++ Entity/Dto/PersonDto.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string first_name { get; set; }
         public string first_last_name { get; set; }
-        public int phone_number { get; set; }
-        public int number_identification { get; set; }
+        public long phone_number { get; set; }
+        public long number_identification { get; set; }
     }
 }
